Add explicite dependency on PHP 5.5 to composer.json
This adds sereral keys to the composer.json declaration I find relevant.
The "composer/installers" dependency is discussed in I615c88d. I'm not
absolutely sure if it's a good idea to add it everywhere. The idea is
that it should be added to all repositories with
"type": "mediawiki-extension".

The PSR1 sniff got lost in the previous patch I2322b25 that switched
this code repository to the Wikibase code sniffer rule set.

Change-Id: Id59059c89fea9c83ac7c76e05b120a57e1368883