Replace custom Wikibase with upstream MediaWiki CodeSniffer
The Wikibase team is not actively working on this codebase any more.
The benefits of having a custom CodeSniffer rule set are not needed
any more. Going back to the upstream rule set makes it easier for
others to continue working on this codebase.

Other changes in this patch:
* Absolute class names in @covers tags make them more compatible with
certain tools.
* Listing the PHP version does not provide much benefit in addition to
the PHP version required for MediaWiki core. The line gets outdated
way to eqasily.
* I tried to find "InterlanguageExtension rev 114818" but wasn't able
to.

Change-Id: Icca19aeb25fe41309ba4a58d9e1af4a3ba63066b