Use LanguageLinksHook to sort interwiki links
Using $wgHooks to make sure hooks are the last to run
no longer works, because the new hook system runs them
as the first ones rather than the last (as the old hooks
system did). That is a problem, because Wikibase
uses ContentAlterParserOutput to add its own interwiki links,
same hook as InterwikiSorting used prior to this commit,
and sorting interwiki links must naturally happen after
links are added by Wikibase.


This commit changes the used hook to LanguageLinks,
which seems to suit the purpose better, and avoides the issue
(for now) by separating iw sorting from adding new links by Wikibase,
which still happens in ContentAlterParserOutput.

The same problem can recur, however, should anyone
use LanguageLinks to add new links. A proper permanent
solution would be to have a hook only for sorting,
or an order-dependent hook system. I decided to just
change the hook for now, to get this fixed - we can discuss
a more proper solution afterwards.

Bug: T257625
Change-Id: I100daba7e3b17e42e0a863d2f879d9d89c525476